Terraform/python/requirements.txt
openai
psycopg2-binary
boto3
